Finalize shadow and border styles
