--- Sources/ProcessOutCoreUI/Sources/Core/UIKit+Extensions/NSLayoutConstraint+Extensions.swift
@@ -1,16 +1,0 @@
-//
-//  NSLayoutConstraint+Extensions.swift
-//  ProcessOutCoreUI
-//
-//  Created by Andrii Vysotskyi on 05.12.2022.
-//
-
-import UIKit
-
-extension NSLayoutConstraint {
-
-    func with(priority: UILayoutPriority) -> NSLayoutConstraint {
-        self.priority = priority
-        return self
-    }
-}
